% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-reto_epi.R
\docType{data}
\name{n_muertes}
\alias{n_muertes}
\title{referencia
https://dplyr.tidyverse.org/articles/rowwise.html
reto:
calcular la tasa de casos fatales
fallecidos / total de casos}
\format{
An object of class \code{integer} of length 1.
}
\usage{
n_muertes
}
\description{
referencia
https://dplyr.tidyverse.org/articles/rowwise.html
reto:
calcular la tasa de casos fatales
fallecidos / total de casos
}
\keyword{datasets}
